test: Split browser test-utils
