+++ tests/browser/test-utils.ts
@@ -1,5 +1,3 @@
-import { userEvent } from 'vitest/browser';
-
 import type {
   PartialAdvancedCameraCardConfig,
   RawAdvancedCameraCardConfig,
@@ -8,10 +6,10 @@ import type { Entity } from '../../src/ha/registry/entity/types';
 import type { MediaLoadedInfoEventDetail } from '../../src/types';
 import { createLogAction } from '../../src/utils/action';
 import { isTruthy } from '../../src/utils/basic';
+import { clickElement, deepQuery, deepQueryAll } from './dom';
 import { FakeHASS, type FakeEntityOptions } from './fake-hass';
 import { createFixtureURL, SNAPSHOT_FIXTURE_FILENAME } from './fixtures';
 import type { MountedCard } from './mounted-card';
-import { createTestMediaURL } from './test-media';
 
 export const CAMERA_ENTITY = 'camera.office';
 
@@ -40,39 +38,6 @@ export const createStillImageCameraConfig = (
     refresh_seconds: 0,
   },
 });
-
-const HTTP_NOT_FOUND = 404;
-const HTTP_OK = 200;
-
-/**
- * A media URL that fails the given number of times and then works from there
- * on, so a test can make a camera recover rather than only fail.
- */
-export const createTemporarilyFailingMediaURL = (
-  failures: number,
-  filename?: string,
-): string =>
-  createTestMediaURL([...Array(failures).fill(HTTP_NOT_FOUND), HTTP_OK], true, filename);
-
-/**
- * A media URL that never works, for a camera that is simply broken.
- */
-export const createFailingMediaURL = (): string =>
-  createTestMediaURL([HTTP_NOT_FOUND], true);
-
-/**
- * A media URL that is never answered, for a camera that accepts the request and
- * then says nothing. Silence is a different failure from a refusal, and the
- * only one that can run a loading timeout out.
- */
-export const createUnansweredMediaURL = (): string => createTestMediaURL([]);
-
-/**
- * A media URL that answers once and is then never answered again, for a camera
- * that delivers a picture and goes quiet behind it.
- */
-export const createStallingMediaURL = (filename?: string): string =>
-  createTestMediaURL([HTTP_OK], false, filename);
 
 export interface FakeCameraDescription {
   entityID: string;
@@ -166,60 +131,6 @@ export const createInitializedAutomation = (): RawAdvancedCameraCardConfig => ({
   triggers: [{ trigger: 'initialized' }],
   actions: [createLogAction(CARD_INITIALIZED_MESSAGE.source)],
 });
-
-// `querySelectorAll` does not look inside a shadow root, so a full search has
-// to step through them a level at a time. The node's own root counts because a
-// Lit element renders into that, not into its children.
-const getImmediateShadowRoots = (root: ParentNode): ShadowRoot[] => {
-  const roots = root instanceof Element && root.shadowRoot ? [root.shadowRoot] : [];
-  for (const child of root.querySelectorAll('*')) {
-    if (child.shadowRoot) {
-      roots.push(child.shadowRoot);
-    }
-  }
-  return roots;
-};
-
-/**
- * Get every shadow root at or below an element.
- */
-export const getAllShadowRoots = (root: ParentNode): ShadowRoot[] =>
-  getImmediateShadowRoots(root).flatMap((child) => [child, ...getAllShadowRoots(child)]);
-
-/**
- * Search an element and every shadow root beneath it. The card nests its own
- * components several roots deep, and neither the source tree nor the node
- * suite has a helper for this.
- */
-export const deepQuery = <T extends Element = Element>(
-  root: ParentNode,
-  selector: string,
-): T | null => {
-  const direct = root.querySelector<T>(selector);
-  if (direct) {
-    return direct;
-  }
-  for (const child of getImmediateShadowRoots(root)) {
-    const found = deepQuery<T>(child, selector);
-    if (found) {
-      return found;
-    }
-  }
-  return null;
-};
-
-/**
- * Every match for a selector across an element and the shadow roots beneath it,
- * for asking how many of something the card rendered rather than whether it
- * rendered any.
- */
-export const deepQueryAll = <T extends Element = Element>(
-  root: ParentNode,
-  selector: string,
-): T[] => [
-  ...root.querySelectorAll<T>(selector),
-  ...getImmediateShadowRoots(root).flatMap((child) => deepQueryAll<T>(child, selector)),
-];
 
 export const isMediaLoadedInfoEventDetail = (
   detail: unknown,
@@ -388,54 +299,3 @@ export const getStatusBarStrings = (root: ParentNode): string[] =>
  */
 export const getStatusBarItem = (root: ParentNode, title: string): Element | null =>
   getStatusBarItems(root).find((item) => item.getAttribute('title') === title) ?? null;
-
-// `userEvent.keyboard` is given one string naming every key to press, in which
-// a name of more than one character is wrapped in braces (`{Escape}`) and a
-// single character stands for itself.
-// See: https://vitest.dev/guide/browser/interactivity-api.html#userevent-keyboard
-const asKeyboardInput = (key: string): string => (key.length === 1 ? key : `{${key}}`);
-
-export const pressKey = async (key: string): Promise<void> =>
-  await userEvent.keyboard(asKeyboardInput(key));
-
-export const holdKey = async (key: string): Promise<void> =>
-  await userEvent.keyboard(`{${key}>}`);
-
-export const releaseKey = async (key: string): Promise<void> =>
-  await userEvent.keyboard(`{/${key}}`);
-
-export const pressTab = async (): Promise<void> => await userEvent.tab();
-
-/**
- * Click an element with a real pointer, which is the only kind that carries the
- * browser's own behaviour: the press moves focus, and an element that stops the
- * press doing so leaves it where it was.
- */
-export const clickElement = async (element: Element): Promise<void> =>
-  await userEvent.click(element);
-
-/**
- * Send a `pointerdown` to an element without moving a real pointer, so the page
- * stays scrolled where the test left it: `clickElement` scrolls its target into
- * view before pressing it.
- *
- * The browser does nothing of its own with a press it did not itself deliver,
- * so what follows is only what the card's own listener does.
- */
-export const dispatchPointerDown = (element: Element): void => {
-  element.dispatchEvent(
-    new PointerEvent('pointerdown', { bubbles: true, composed: true }),
-  );
-};
-
-/**
- * The element that actually has focus. `document.activeElement` names the
- * outermost shadow host in the way, since focus is reported per tree.
- */
-export const getFocusedElement = (): Element | null => {
-  let focused = document.activeElement;
-  while (focused?.shadowRoot?.activeElement) {
-    focused = focused.shadowRoot.activeElement;
-  }
-  return focused;
-};
